+++ words/cli/param_types.py
@@ -3,6 +3,7 @@
 from click import ParamType
 
 from words import WordsError, bp  # noqa
+from words.cli import ui  # noqa
 
 
 class RangeType(ParamType):